backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-pptx>=0.6.21
pillow>=11.0.0
requests>=2.31.0
openai>=1.14.0
pydantic>=2.8.0
boto3>=1.34.0
python-dotenv>=1.0.0
gtts>=2.4.0
python-multipart>=0.0.6
